label: don't stub functions that are already stubbed
the selinux package already has stubs for most of these,
so instead of maintaining separate stubs, use the aliases.
